Diff Foreign Key
Add an option to the diff function so that when generating a diff on a field that is the foreign key, get what the string representation of the foreign object. For example, if the field is a FK to a User, it will display the username instead of the id.